requirements.txt
beautifulsoup4==4.9.1
bs4==0.0.1
cached-property==1.5.1
certifi==2020.6.20
cffi==1.14.2
chardet==3.0.4
commonmark==0.9.1
cryptography==3.1
decorator==4.4.2
dictdiffer==0.8.1
idna==2.10
notion==0.0.25
pycparser==2.20
python-slugify==4.0.1
python-telegram-bot==12.8
pytz==2020.1
requests==2.24.0
six==1.15.0
soupsieve==2.0.1
text-unidecode==1.3
tornado==6.0.4
tzlocal==2.1
urllib3==1.25.10
